implemented services for finding offer price, created new route for grand total of each customers

+++ models/products.js
@@ -11,7 +11,7 @@ const Product = sequelize.define('product',{
         field: "id"
     },
     productName: DataTypes.STRING,
-    price: DataTypes.INTEGER,
-    offerPrice:DataTypes.INTEGER
+    price: DataTypes.FLOAT,
+    offerPrice:DataTypes.FLOAT
 })
 module.exports = Product;
